% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcap_to_zeek.R
\name{find_zeek}
\alias{find_zeek}
\title{Find the Zeek binary}
\usage{
find_zeek(path = Sys.getenv("ZEEK_PATH", ""))
}
\arguments{
\item{path}{hint to where to look for the Zeek binary}
}
\value{
length 1 character vector of the path to the zeek binary or \code{""}
}
\description{
Use the environment variable \code{ZEEK_PATH} or specify the directory in
the call to this function.
}
\examples{
find_zeek()
}
