% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-zeek-logs.R
\name{read_zeek_logs}
\alias{read_zeek_logs}
\title{Read zeek logs from a processed PCAP into a list}
\usage{
read_zeek_logs(log_dir)
}
\arguments{
\item{log_dir}{directory of zeek logs}
}
\description{
Read zeek logs from a processed PCAP into a list
}
\note{
Logs must be in Parquet or JSON format.
}
\examples{
loc <- tryCatch(
  pcap_to_zeek(system.file("pcap/ssh.pcap", package = "zeekr")),
  error = function(e) message("No Zeek")
)

if (!is.null(loc)) {
  read_zeek_logs(loc)
  unlink(loc) # don't do this IRL until you're done working with or saving.
}
}
