% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeek.R
\name{zeek}
\alias{zeek}
\title{Call the Zeek binary with optional custom environment variables and options}
\usage{
zeek(zeek_bin = find_zeek(), args = c(), env = c())
}
\arguments{
\item{zeek_bin}{specify a complate path or let \code{\link[=find_zeek]{find_zeek()}} do the dirty work.}

\item{args}{same as \code{\link[=system2]{system2()}} \code{args}}

\item{env}{same as \code{\link[=system2]{system2()}} \code{env}}
}
\value{
\code{list} with \code{stderr}, \code{stdout}, \code{status} and \code{errmsg} (invisibly)
}
\description{
This is just a convenience wrapper around \code{\link[=system2]{system2()}}. See \code{\link[=find_zeek]{find_zeek()}} for
information on helping this package find the Zeek binary.
}
