% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeekr-package.R
\docType{package}
\name{zeekr}
\alias{zeekr}
\alias{zeekr-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/zeekr}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/zeekr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
