% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeekr-package.R
\docType{package}
\name{zeekr}
\alias{zeekr}
\alias{zeekr-package}
\title{Tools to Make Analyses Using Zeek Easier}
\description{
\href{https://zeek.org/}{Zeek} is an open source network security monitoring
system. Tools are provided to make it a bit easier to work with Zeek to perform
analyses with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/zeekr}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/zeekr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
